Rename model to entities
