+++ api/services/annotations.go
@@ -40,15 +40,15 @@ import (
 	"time"
 
 	"github.com/ausocean/openfish/api/ds_client"
-	"github.com/ausocean/openfish/api/model"
+	"github.com/ausocean/openfish/api/entities"
 	"github.com/ausocean/openfish/datastore"
 )
 
 // GetAnnotationByID gets an annotation from datastore when provided with an ID.
-func GetAnnotationByID(id int64) (*model.Annotation, error) {
+func GetAnnotationByID(id int64) (*entities.Annotation, error) {
 	store := ds_client.Get()
-	key := store.IDKey(model.ANNOTATION_KIND, id)
-	var annotation model.Annotation
+	key := store.IDKey(entities.ANNOTATION_KIND, id)
+	var annotation entities.Annotation
 	err := store.Get(context.Background(), key, &annotation)
 	if err != nil {
 		return nil, err
@@ -59,17 +59,17 @@ func GetAnnotationByID(id int64) (*model.Annotation, error) {
 
 func AnnotationExists(id int64) bool {
 	store := ds_client.Get()
-	key := store.IDKey(model.ANNOTATION_KIND, id)
-	var annotation model.Annotation
+	key := store.IDKey(entities.ANNOTATION_KIND, id)
+	var annotation entities.Annotation
 	err := store.Get(context.Background(), key, &annotation)
 	return err == nil
 }
 
 // GetAnnotations gets a list of annotations, filtering by timespan, capturesource, observer & observation if specified.
-func GetAnnotations(limit int, offset int, observer *string, observation map[string]string) ([]model.Annotation, []int64, error) {
+func GetAnnotations(limit int, offset int, observer *string, observation map[string]string) ([]entities.Annotation, []int64, error) {
 	// Fetch data from the datastore.
 	store := ds_client.Get()
-	query := store.NewQuery(model.ANNOTATION_KIND, false)
+	query := store.NewQuery(entities.ANNOTATION_KIND, false)
 
 	// Filter by observer.
 	if observer != nil {
@@ -88,10 +88,10 @@ func GetAnnotations(limit int, offset int, observer *string, observation map[str
 	query.Limit(limit)
 	query.Offset(offset)
 
-	var annotations []model.Annotation
+	var annotations []entities.Annotation
 	keys, err := store.GetAll(context.Background(), query, &annotations)
 	if err != nil {
-		return []model.Annotation{}, []int64{}, err
+		return []entities.Annotation{}, []int64{}, err
 	}
 	ids := make([]int64, len(annotations))
 	for i, k := range keys {
@@ -102,7 +102,7 @@ func GetAnnotations(limit int, offset int, observer *string, observation map[str
 }
 
 // CreateAnnotation creates a new annotation.
-func CreateAnnotation(videoStreamID int64, timeSpan model.TimeSpan, boundingBox *model.BoundingBox, observer string, observation map[string]string) (int64, error) {
+func CreateAnnotation(videoStreamID int64, timeSpan entities.TimeSpan, boundingBox *entities.BoundingBox, observer string, observation map[string]string) (int64, error) {
 	// Convert observation map into a format the datastore can take.
 	obsKeys := make([]string, 0, len(observation))
 	obsPairs := make([]string, 0, len(observation))
@@ -113,7 +113,7 @@ func CreateAnnotation(videoStreamID int64, timeSpan model.TimeSpan, boundingBox 
 	}
 
 	// Create annotation entity and add to the datastore.
-	an := model.Annotation{
+	an := entities.Annotation{
 		VideoStreamID:    videoStreamID,
 		TimeSpan:         timeSpan,
 		BoundingBox:      boundingBox,
@@ -129,7 +129,7 @@ func CreateAnnotation(videoStreamID int64, timeSpan model.TimeSpan, boundingBox 
 
 	// Get a unique ID for the new annotation.
 	store := ds_client.Get()
-	key := store.IncompleteKey(model.ANNOTATION_KIND)
+	key := store.IncompleteKey(entities.ANNOTATION_KIND)
 	key, err := store.Put(context.Background(), key, &an)
 	if err != nil {
 		return 0, err
@@ -144,11 +144,11 @@ func UpdateAnnotation(id int64, streamURL *string, captureSource *int64, startTi
 
 	// Update data in the datastore.
 	store := ds_client.Get()
-	key := store.IDKey(model.VIDEOSTREAM_KIND, id)
-	var videoStream model.VideoStream
+	key := store.IDKey(entities.VIDEOSTREAM_KIND, id)
+	var videoStream entities.VideoStream
 
 	return store.Update(context.Background(), key, func(e datastore.Entity) {
-		v, ok := e.(*model.VideoStream)
+		v, ok := e.(*entities.VideoStream)
 		if ok {
 			if streamURL != nil {
 				v.StreamUrl = *streamURL
@@ -173,6 +173,6 @@ func DeleteAnnotation(id int64) error {
 
 	// Delete entity.
 	store := ds_client.Get()
-	key := store.IDKey(model.ANNOTATION_KIND, id)
+	key := store.IDKey(entities.ANNOTATION_KIND, id)
 	return store.Delete(context.Background(), key)
 }
