+++ api/services/videostream_test.go
@@ -37,7 +37,7 @@ import (
 	"testing"
 	"time"
 
-	"github.com/ausocean/openfish/api/model"
+	"github.com/ausocean/openfish/api/entities"
 	"github.com/ausocean/openfish/api/services"
 )
 
@@ -205,7 +205,7 @@ func TestDeleteVideoStreamWithAssociatedAnnotations(t *testing.T) {
 	cs, _ := services.CreateCaptureSource("Stony Point camera 1", 0.0, 0.0, "RPI camera", nil)
 	id, _ := services.CreateVideoStream("http://youtube.com/watch?v=abc123", int64(cs), _8am, &_4pm)
 	services.CreateAnnotation(id,
-		model.TimeSpan{Start: _8am, End: _9am},
+		entities.TimeSpan{Start: _8am, End: _9am},
 		nil, "scott@ausocean.org",
 		map[string]string{"species": "Sepia Apama"})
